server/requirements.txt
mysql-connector-python>=9.0.0
flask>=3.0.3